+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/Config.java
@@ -189,6 +189,8 @@ public interface Config {
         long getCreativeRangeUpgradeEnergyUsage();
 
         int getRangeUpgradeRange();
+
+        long getAutocraftingUpgradeEnergyUsage();
     }
 
     interface WirelessGridEntry {