+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/exporter/ExporterContainerMenu.java
@@ -8,37 +8,45 @@ import com.refinedmods.refinedstorage.common.support.containermenu.AbstractSimpl
 import com.refinedmods.refinedstorage.common.support.containermenu.ClientProperty;
 import com.refinedmods.refinedstorage.common.support.containermenu.PropertyTypes;
 import com.refinedmods.refinedstorage.common.support.containermenu.ServerProperty;
-import com.refinedmods.refinedstorage.common.support.resource.ResourceContainerData;
+import com.refinedmods.refinedstorage.common.support.exportingindicator.ExportingIndicator;
+import com.refinedmods.refinedstorage.common.support.exportingindicator.ExportingIndicatorListener;
+import com.refinedmods.refinedstorage.common.support.exportingindicator.ExportingIndicators;
 import com.refinedmods.refinedstorage.common.upgrade.UpgradeContainer;
 import com.refinedmods.refinedstorage.common.upgrade.UpgradeDestinations;
 
 import net.minecraft.network.chat.MutableComponent;
+import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.entity.player.Player;
 
 import static com.refinedmods.refinedstorage.common.util.IdentifierUtil.createTranslation;
 
-public class ExporterContainerMenu extends AbstractSimpleFilterContainerMenu<AbstractExporterBlockEntity> {
+public class ExporterContainerMenu extends AbstractSimpleFilterContainerMenu<AbstractExporterBlockEntity>
+    implements ExportingIndicatorListener {
     private static final MutableComponent FILTER_HELP = createTranslation("gui", "exporter.filter_help");
+
+    private final ExportingIndicators indicators;
 
     public ExporterContainerMenu(final int syncId,
                                  final Inventory playerInventory,
-                                 final ResourceContainerData resourceContainerData) {
+                                 final ExporterData data) {
         super(
             Menus.INSTANCE.getExporter(),
             syncId,
             playerInventory.player,
-            resourceContainerData,
+            data.resourceContainerData(),
             UpgradeDestinations.EXPORTER,
             FILTER_HELP
         );
+        this.indicators = new ExportingIndicators(data.exportingIndicators());
     }
 
     ExporterContainerMenu(final int syncId,
                           final Player player,
                           final AbstractExporterBlockEntity exporter,
                           final ResourceContainer resourceContainer,
-                          final UpgradeContainer upgradeContainer) {
+                          final UpgradeContainer upgradeContainer,
+                          final ExportingIndicators indicators) {
         super(
             Menus.INSTANCE.getExporter(),
             syncId,
@@ -48,6 +56,23 @@ public class ExporterContainerMenu extends AbstractSimpleFilterContainerMenu<Abs
             exporter,
             FILTER_HELP
         );
+        this.indicators = indicators;
+    }
+
+    ExportingIndicator getIndicator(final int idx) {
+        return indicators.get(idx);
+    }
+
+    int getIndicators() {
+        return indicators.size();
+    }
+
+    @Override
+    public void broadcastChanges() {
+        super.broadcastChanges();
+        if (player instanceof ServerPlayer serverPlayer) {
+            indicators.detectChanges(serverPlayer);
+        }
     }
 
     @Override
@@ -74,5 +99,10 @@ public class ExporterContainerMenu extends AbstractSimpleFilterContainerMenu<Abs
             blockEntity::getSchedulingModeType,
             blockEntity::setSchedulingModeType
         ));
+    }
+
+    @Override
+    public void indicatorChanged(final int index, final ExportingIndicator indicator) {
+        indicators.set(index, indicator);
     }
 }