+++ refinedstorage-fabric/src/main/java/com/refinedmods/refinedstorage/fabric/importer/FabricStorageImporterTransferStrategyFactory.java
@@ -1,14 +1,12 @@
 package com.refinedmods.refinedstorage.fabric.importer;
 
 import com.refinedmods.refinedstorage.api.core.NullableType;
-import com.refinedmods.refinedstorage.api.network.impl.node.importer.ImporterSource;
 import com.refinedmods.refinedstorage.api.network.impl.node.importer.ImporterTransferStrategyImpl;
 import com.refinedmods.refinedstorage.api.network.node.importer.ImporterTransferStrategy;
 import com.refinedmods.refinedstorage.api.resource.ResourceKey;
 import com.refinedmods.refinedstorage.common.api.importer.ImporterTransferStrategyFactory;
-import com.refinedmods.refinedstorage.common.api.support.network.AmountOverride;
 import com.refinedmods.refinedstorage.common.api.upgrade.UpgradeState;
-import com.refinedmods.refinedstorage.common.content.Items;
+import com.refinedmods.refinedstorage.common.importer.ImporterTransferQuotaProvider;
 
 import java.util.function.Function;
 
@@ -38,20 +36,20 @@ public class FabricStorageImporterTransferStrategyFactory<P> implements Importer
     public ImporterTransferStrategy create(final ServerLevel level,
                                            final BlockPos pos,
                                            final Direction direction,
-                                           final UpgradeState upgradeState,
-                                           final AmountOverride amountOverride) {
-        final ImporterSource source = new FabricStorageImporterSource<>(
+                                           final UpgradeState upgradeState) {
+        final FabricStorageImporterSource<P> source = new FabricStorageImporterSource<>(
             lookup,
             fromPlatformMapper,
             toPlatformMapper,
             level,
             pos,
-            direction,
-            amountOverride
+            direction
         );
-        final long transferQuota = upgradeState.has(Items.INSTANCE.getStackUpgrade())
-            ? singleAmount * 64
-            : singleAmount;
-        return new ImporterTransferStrategyImpl(source, transferQuota);
+        final ImporterTransferQuotaProvider transferQuotaProvider = new ImporterTransferQuotaProvider(
+            singleAmount,
+            upgradeState,
+            source::getAmount
+        );
+        return new ImporterTransferStrategyImpl(source, transferQuotaProvider);
     }
 }