+++ refinedstorage-fabric/src/main/java/com/refinedmods/refinedstorage/fabric/storage/FabricStorageInsertableStorage.java
@@ -5,7 +5,6 @@ import com.refinedmods.refinedstorage.api.core.NullableType;
 import com.refinedmods.refinedstorage.api.resource.ResourceKey;
 import com.refinedmods.refinedstorage.api.storage.Actor;
 import com.refinedmods.refinedstorage.api.storage.InsertableStorage;
-import com.refinedmods.refinedstorage.common.api.support.network.AmountOverride;
 
 import java.util.function.Function;
 
@@ -21,18 +20,27 @@ public class FabricStorageInsertableStorage<T> implements InsertableStorage {
     private final BlockApiCache<Storage<T>, Direction> cache;
     private final Function<ResourceKey, @NullableType T> toPlatformMapper;
     private final Direction direction;
-    private final AmountOverride amountOverride;
 
     public FabricStorageInsertableStorage(final BlockApiLookup<Storage<T>, Direction> lookup,
                                           final Function<ResourceKey, @NullableType T> toPlatformMapper,
                                           final ServerLevel serverLevel,
                                           final BlockPos pos,
-                                          final Direction direction,
-                                          final AmountOverride amountOverride) {
+                                          final Direction direction) {
         this.cache = BlockApiCache.create(lookup, serverLevel, pos);
         this.toPlatformMapper = toPlatformMapper;
         this.direction = direction;
-        this.amountOverride = amountOverride;
+    }
+
+    public long getAmount(final ResourceKey resource) {
+        final Storage<T> storage = cache.find(direction);
+        if (storage == null) {
+            return 0;
+        }
+        final T platformResource = toPlatformMapper.apply(resource);
+        if (platformResource == null) {
+            return 0;
+        }
+        return FabricStorageUtil.getCurrentAmount(storage, platformResource);
     }
 
     @Override
@@ -45,15 +53,7 @@ public class FabricStorageInsertableStorage<T> implements InsertableStorage {
         if (platformResource == null) {
             return 0;
         }
-        final long correctedAmount = amountOverride.overrideAmount(
-            resource,
-            amount,
-            () -> FabricStorageUtil.getCurrentAmount(storage, platformResource)
-        );
-        if (correctedAmount == 0) {
-            return 0;
-        }
-        return doInsert(platformResource, correctedAmount, action, storage);
+        return doInsert(platformResource, amount, action, storage);
     }
 
     private long doInsert(final T platformResource, final long amount, final Action action, final Storage<T> storage) {