+++ refinedstorage-neoforge/src/main/java/com/refinedmods/refinedstorage/neoforge/exporter/FluidHandlerExporterTransferStrategyFactory.java
@@ -1,39 +1,61 @@
 package com.refinedmods.refinedstorage.neoforge.exporter;
 
 import com.refinedmods.refinedstorage.api.network.impl.node.exporter.ExporterTransferStrategyImpl;
+import com.refinedmods.refinedstorage.api.network.impl.node.exporter.MissingResourcesListeningExporterTransferStrategy;
 import com.refinedmods.refinedstorage.api.network.node.exporter.ExporterTransferStrategy;
+import com.refinedmods.refinedstorage.api.resource.ResourceKey;
 import com.refinedmods.refinedstorage.common.Platform;
 import com.refinedmods.refinedstorage.common.api.exporter.ExporterTransferStrategyFactory;
-import com.refinedmods.refinedstorage.common.api.support.network.AmountOverride;
+import com.refinedmods.refinedstorage.common.api.storage.root.FuzzyRootStorage;
 import com.refinedmods.refinedstorage.common.api.upgrade.UpgradeState;
 import com.refinedmods.refinedstorage.common.content.Items;
-import com.refinedmods.refinedstorage.common.exporter.FuzzyExporterTransferStrategy;
+import com.refinedmods.refinedstorage.common.exporter.ExporterTransferQuotaProvider;
+import com.refinedmods.refinedstorage.common.support.resource.FluidResource;
 import com.refinedmods.refinedstorage.neoforge.storage.CapabilityCache;
 import com.refinedmods.refinedstorage.neoforge.storage.CapabilityCacheImpl;
 import com.refinedmods.refinedstorage.neoforge.storage.FluidHandlerInsertableStorage;
+
+import java.util.function.ToLongFunction;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.server.level.ServerLevel;
 
+import static com.refinedmods.refinedstorage.api.network.impl.node.exporter.MissingResourcesListeningExporterTransferStrategy.OnMissingResources.scheduleAutocrafting;
+
 public class FluidHandlerExporterTransferStrategyFactory implements ExporterTransferStrategyFactory {
+    @Override
+    public Class<? extends ResourceKey> getResourceType() {
+        return FluidResource.class;
+    }
+
     @Override
     public ExporterTransferStrategy create(final ServerLevel level,
                                            final BlockPos pos,
                                            final Direction direction,
                                            final UpgradeState upgradeState,
-                                           final AmountOverride amountOverride,
                                            final boolean fuzzyMode) {
         final CapabilityCache coordinates = new CapabilityCacheImpl(level, pos, direction);
-        final FluidHandlerInsertableStorage destination = new FluidHandlerInsertableStorage(
-            coordinates,
-            amountOverride
+        final FluidHandlerInsertableStorage destination = new FluidHandlerInsertableStorage(coordinates);
+        final long singleAmount = Platform.INSTANCE.getBucketAmount();
+        final ExporterTransferStrategy strategy = create(
+            fuzzyMode,
+            destination,
+            new ExporterTransferQuotaProvider(singleAmount, upgradeState, destination::getAmount, true)
         );
-        final long transferQuota = (upgradeState.has(Items.INSTANCE.getStackUpgrade()) ? 64 : 1)
-            * Platform.INSTANCE.getBucketAmount();
-        if (fuzzyMode) {
-            return new FuzzyExporterTransferStrategy(destination, transferQuota);
+        if (upgradeState.has(Items.INSTANCE.getAutocraftingUpgrade())) {
+            return new MissingResourcesListeningExporterTransferStrategy(strategy, scheduleAutocrafting(
+                new ExporterTransferQuotaProvider(singleAmount, upgradeState, destination::getAmount, false)));
         }
-        return new ExporterTransferStrategyImpl(destination, transferQuota);
+        return strategy;
+    }
+
+    private ExporterTransferStrategy create(final boolean fuzzyMode,
+                                            final FluidHandlerInsertableStorage destination,
+                                            final ToLongFunction<ResourceKey> transferQuotaProvider) {
+        if (fuzzyMode) {
+            return new ExporterTransferStrategyImpl(destination, transferQuotaProvider, FuzzyRootStorage.expander());
+        }
+        return new ExporterTransferStrategyImpl(destination, transferQuotaProvider);
     }
 }