+++ refinedstorage-neoforge/src/main/java/com/refinedmods/refinedstorage/neoforge/importer/FluidHandlerImporterTransferStrategyFactory.java
@@ -1,12 +1,11 @@
 package com.refinedmods.refinedstorage.neoforge.importer;
 
-import com.refinedmods.refinedstorage.api.network.impl.node.importer.ImporterSource;
 import com.refinedmods.refinedstorage.api.network.impl.node.importer.ImporterTransferStrategyImpl;
 import com.refinedmods.refinedstorage.api.network.node.importer.ImporterTransferStrategy;
 import com.refinedmods.refinedstorage.common.api.importer.ImporterTransferStrategyFactory;
-import com.refinedmods.refinedstorage.common.api.support.network.AmountOverride;
 import com.refinedmods.refinedstorage.common.api.upgrade.UpgradeState;
 import com.refinedmods.refinedstorage.common.content.Items;
+import com.refinedmods.refinedstorage.common.importer.ImporterTransferQuotaProvider;
 import com.refinedmods.refinedstorage.neoforge.storage.CapabilityCacheImpl;
 
 import net.minecraft.core.BlockPos;
@@ -19,16 +18,20 @@ public class FluidHandlerImporterTransferStrategyFactory implements ImporterTran
     public ImporterTransferStrategy create(final ServerLevel level,
                                            final BlockPos pos,
                                            final Direction direction,
-                                           final UpgradeState upgradeState,
-                                           final AmountOverride amountOverride) {
-        final ImporterSource source = new FluidHandlerImporterSource(new CapabilityCacheImpl(
+                                           final UpgradeState upgradeState) {
+        final FluidHandlerImporterSource source = new FluidHandlerImporterSource(new CapabilityCacheImpl(
             level,
             pos,
             direction
-        ), amountOverride);
-        final int transferQuota = upgradeState.has(Items.INSTANCE.getStackUpgrade())
+        ));
+        final int singleAmount = upgradeState.has(Items.INSTANCE.getStackUpgrade())
             ? FluidType.BUCKET_VOLUME * 64
             : FluidType.BUCKET_VOLUME;
-        return new ImporterTransferStrategyImpl(source, transferQuota);
+        final ImporterTransferQuotaProvider transferQuotaProvider = new ImporterTransferQuotaProvider(
+            singleAmount,
+            upgradeState,
+            source::getAmount
+        );
+        return new ImporterTransferStrategyImpl(source, transferQuotaProvider);
     }
 }