+++ CHANGELOG.md
@@ -7,6 +7,14 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+-   Autocrafting Upgrade
+
+### Changed
+
+-   The filter slots for the Exporter now display whether a resource is missing, the destination does not accept it, the resource cannot be autocrafted due to missing resources, or whether the resource is currently autocrafting.
+
 ## [2.0.0-milestone.4.13] - 2025-02-01
 
 ### Added